Change project withi pyproject & pre-commit, change structure fo project #85
