--- environment_linux.yml
@@ -1,148 +1,0 @@
-name: vcub_keeper
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - argon2-cffi=20.1.0=py38h7b6447c_1
-  - backcall=0.2.0=py_0
-  - ca-certificates=2020.10.14=0
-  - certifi=2020.6.20=py38h06a4308_2
-  - decorator=4.4.2=py_0
-  - defusedxml=0.6.0=py_0
-  - entrypoints=0.3=py38_0
-  - importlib_metadata=1.7.0=0
-  - ipykernel=5.3.4=py38h5ca1d4c_0
-  - ipython_genutils=0.2.0=py38_0
-  - jedi=0.17.2=py38h32f6830_1
-  - jinja2=2.11.2=py_0
-  - jsonschema=3.2.0=py38_0
-  - jupyter_core=4.6.3=py38_0
-  - ld_impl_linux-64=2.33.1=h53a641e_7
-  - libedit=3.1.20191231=h14c3975_1
-  - libffi=3.3=he6710b0_2
-  - libgcc-ng=9.1.0=hdf63c60_0
-  - libsodium=1.0.18=h7b6447c_0
-  - libstdcxx-ng=9.1.0=hdf63c60_0
-  - markupsafe=1.1.1=py38h7b6447c_0
-  - mistune=0.8.4=py38h7b6447c_1000
-  - nb_conda=2.2.1=py38_1
-  - nb_conda_kernels=2.2.3=py38_0
-  - ncurses=6.2=he6710b0_1
-  - notebook=6.1.1=py38_0
-  - openssl=1.1.1h=h7b6447c_0
-  - packaging=20.4=py_0
-  - pandoc=2.10.1=0
-  - pandocfilters=1.4.2=py38_1
-  - parso=0.7.1=pyh9f0ad1d_0
-  - pexpect=4.8.0=py38_0
-  - pickleshare=0.7.5=py38_1000
-  - pip=20.2.4=py38_0
-  - prometheus_client=0.8.0=py_0
-  - pycparser=2.20=py_2
-  - pyparsing=2.4.7=py_0
-  - pyrsistent=0.16.0=py38h7b6447c_0
-  - python=3.8.5=hcff3b4d_0
-  - python-dateutil=2.8.1=py_0
-  - python_abi=3.8=1_cp38
-  - readline=8.0=h7b6447c_0
-  - send2trash=1.5.0=py38_0
-  - six=1.15.0=py_0
-  - sqlite=3.33.0=h62c20be_0
-  - terminado=0.8.3=py38_0
-  - testpath=0.4.4=py_0
-  - tk=8.6.10=hbc83047_0
-  - tornado=6.0.4=py38h7b6447c_1
-  - wcwidth=0.2.5=py_0
-  - webencodings=0.5.1=py38_1
-  - wheel=0.34.2=py38_0
-  - xz=5.2.5=h7b6447c_0
-  - zeromq=4.3.2=he6710b0_2
-  - zlib=1.2.11=h7b6447c_3
-  - pip:
-    - async-generator==1.10
-    - attrs==20.2.0
-    - bleach==3.2.1
-    - brotli==1.0.9
-    - cffi==1.14.3
-    - chardet==3.0.4
-    - click==7.1.2
-    - click-plugins==1.1.1
-    - cligj==0.7.1
-    - cycler==0.10.0
-    - dash==1.16.3
-    - dash-core-components==1.12.1
-    - dash-html-components==1.1.1
-    - dash-renderer==1.8.2
-    - dash-table==4.10.1
-    - fiona==1.8.19
-    - flask==1.1.2
-    - flask-compress==1.7.0
-    - future==0.18.2
-    - geopandas==0.9.0
-    - gunicorn==20.0.4
-    - idna==2.10
-    - importlib-metadata==2.0.0
-    - ipython==7.18.1
-    - ipywidgets==7.6.3
-    - itsdangerous==1.1.0
-    - joblib==0.17.0
-    - jupyter-client==6.1.7
-    - jupyter-contrib-core==0.3.3
-    - jupyter-contrib-nbextensions==0.5.1
-    - jupyter-highlight-selected-word==0.2.0
-    - jupyter-latex-envs==1.4.6
-    - jupyter-nbextensions-configurator==0.4.1
-    - jupyterlab-pygments==0.1.2
-    - jupyterlab-widgets==1.0.0
-    - jupyterthemes==0.20.0
-    - keplergl==0.2.2
-    - kiwisolver==1.2.0
-    - lesscpy==0.14.0
-    - llvmlite==0.34.0
-    - lsanomaly==0.1.3
-    - lxml==4.6.1
-    - matplotlib==3.3.1
-    - munch==2.5.0
-    - nbclient==0.5.1
-    - nbconvert==6.0.7
-    - nbformat==5.0.8
-    - nest-asyncio==1.4.1
-    - numba==0.51.2
-    - numexpr==2.7.1
-    - numpy==1.19.1
-    - pandas==1.1.1
-    - patsy==0.5.1
-    - pillow==7.2.0
-    - plotly==4.9.0
-    - plotly-express==0.4.1
-    - ply==3.11
-    - prompt-toolkit==3.0.8
-    - ptyprocess==0.6.0
-    - pygments==2.7.1
-    - pyproj==3.0.1
-    - python-dotenv==0.14.0
-    - pytz==2020.1
-    - pyweatherbit==2.0.0
-    - pyyaml==5.3.1
-    - pyzmq==19.0.2
-    - requests==2.24.0
-    - responses==0.12.0
-    - retrying==1.3.3
-    - rrcf==0.4.3
-    - scikit-learn==0.23.2
-    - scipy==1.5.2
-    - seaborn==0.11.0
-    - setuptools==50.3.2
-    - shapely==1.7.1
-    - statsmodels==0.11.1
-    - tables==3.6.1
-    - tabulate==0.8.7
-    - threadpoolctl==2.1.0
-    - tqdm==4.43.0
-    - traitlets==5.0.5
-    - traittypes==0.2.1
-    - urllib3==1.25.10
-    - werkzeug==1.0.1
-    - widgetsnbextension==3.5.1
-    - zipp==3.3.1
