--- requirements.txt
@@ -1,175 +1,0 @@
-altair==4.2.0
-argon2-cffi==20.1.0
-async-generator==1.10
-attrs==20.2.0
-backcall==0.2.0
-backoff==1.11.1
-bleach==3.2.1
-boltons==21.0.0
-Brotli==1.0.9
-catboost==1.0.5
-certifi==2020.6.20
-cffi==1.14.3
-chardet==3.0.4
-click==7.1.2
-click-plugins==1.1.1
-click-spinner==0.1.10
-cligj==0.7.1
-cloudpickle==2.0.0
-colorlog==6.6.0
-coverage==6.2
-cycler==0.10.0
-dacite==1.6.0
-dash==2.3.1
-dash-bootstrap-components==1.1.0
-dash-core-components==2.0.0
-dash-daq==0.5.0
-dash-html-components==2.0.0
-dash-renderer==1.8.3
-dash-table==5.0.0
-datacommons==1.4.3
-datacommons-pandas==0.0.3
-datapane==0.14.0
-decorator==4.4.2
-defusedxml==0.6.0
-dominate==2.6.0
-entrypoints==0.3
-eurybia==1.0.0
-face==20.1.1
-Fiona==1.8.19
-flake8==4.0.1
-Flask==1.1.2
-Flask-Compress==1.7.0
-furl==2.1.3
-future==0.18.2
-geopandas==0.9.0
-glom==20.11.0
-graphviz==0.20
-gunicorn==20.0.4
-idna==2.10
-importlib-metadata==2.0.0
-importlib-resources==5.7.1
-iniconfig==1.1.1
-ipykernel
-ipython==7.18.1
-ipython-genutils==0.2.0
-ipywidgets==7.6.3
-itsdangerous==1.1.0
-jedi==0.17.2
-Jinja2==2.11.2
-joblib==0.17.0
-jsonschema==3.2.0
-jupyter==1.0.0
-jupyter-client
-jupyter-console
-jupyter-contrib-core==0.3.3
-jupyter-contrib-nbextensions==0.5.1
-jupyter-core==4.6.3
-jupyter-highlight-selected-word==0.2.0
-jupyter-latex-envs==1.4.6
-jupyter-nbextensions-configurator==0.4.1
-jupyterlab-pygments==0.1.2
-jupyterlab-widgets==1.0.0
-jupyterthemes==0.20.0
-keplergl==0.2.2
-kiwisolver==1.2.0
-lesscpy==0.14.0
-llvmlite==0.38.0
-lsanomaly==0.1.3
-lxml==4.6.1
-MarkupSafe==1.1.1
-matplotlib==3.3.1
-mccabe==0.6.1
-micawber==0.5.4
-mistune==0.8.4
-monotonic==1.6
-munch==2.5.0
-#nb-conda==2.2.1
-#nb-conda-kernels==2.2.3
-nbclient==0.5.1
-nbconvert==6.0.7
-nbformat==5.0.8
-nest-asyncio==1.4.1
-notebook==6.1.1
-numba==0.55.1
-numexpr==2.7.1
-numpy==1.19.1
-orderedmultidict==1.0.1
-packaging
-pandas==1.1.1
-pandocfilters==1.4.2
-parso==0.7.1
-patsy==0.5.1
-pexpect==4.8.0
-pickleshare==0.7.5
-Pillow==7.2.0
-pip==20.2.4
-plotly==5.7.0
-plotly-express==0.4.1
-pluggy==1.0.0
-ply==3.11
-posthog==1.4.7
-prometheus-client==0.8.0
-prompt-toolkit==3.0.8
-ptyprocess==0.6.0
-py==1.11.0
-pyarrow==6.0.1
-pycodestyle==2.8.0
-pycparser==2.20
-pydantic==1.9.0
-pyflakes==2.4.0
-Pygments==2.7.1
-pyparsing==2.4.7
-pyproj==3.0.1
-pyrsistent==0.16.0
-pytest==6.2.5
-pytest-cov==3.0.0
-python-dateutil==2.8.1
-python-dotenv==0.14.0
-pytz==2020.1
-pyweatherbit==2.0.0
-PyYAML==6.0
-pyzmq==19.0.2
-qtconsole==5.3.0
-QtPy==2.1.0
-requests==2.28.0
-requests-toolbelt==0.9.1
-responses==0.12.0
-retrying==1.3.3
-rrcf==0.4.3
-scikit-learn==0.24.2
-scipy==1.5.2
-seaborn==0.11.0
-Send2Trash==1.5.0
-setuptools==61.3.0
-shap==0.40.0
-shapash==2.0.0
-Shapely==1.7.1
-six==1.15.0
-slicer==0.0.7
-statsmodels==0.11.1
-stringcase==1.2.0
-tables==3.6.1
-tabulate==0.8.7
-tenacity==8.0.1
-terminado==0.8.3
-testpath==0.4.4
-threadpoolctl==2.1.0
-toml==0.10.2
-tomli==2.0.0
-toolz==0.11.2
-tornado==6.0.4
-tqdm==4.43.0
-traitlets==5.0.5
-traittypes==0.2.1
-typing-extensions==4.2.0
-urllib3==1.25.10
-validators==0.18.2
-#vcub-keeper==1.2
-vega-datasets==0.9.0
-wcwidth==0.2.5
-webencodings==0.5.1
-Werkzeug==1.0.1
-wheel==0.34.2
-widgetsnbextension==3.5.1
-zipp==3.3.1
