+++ src/vcub_keeper/__init__.py
@@ -1,0 +1,1 @@
+__version__ = '1.2.1'
