+++ src/vcub_keeper/test.py
@@ -1,0 +1,27 @@
+# import os
+
+# ROOT_DIR = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
+# print("ROOT_DIR:", ROOT_DIR)
+
+# ROOT_DATA_RAW = ROOT_DIR + "/data/raw/"
+# print("ROOT_DATA_RAW:", ROOT_DATA_RAW)
+
+from vcub_keeper.config import ROOT_DIR  # , ROOT_DATA_RAW, ROOT_DATA_CLEAN, ROOT_DATA_REF, ROOT_MODEL, ROOT_TESTS_DATA
+
+print("ROOT_DIR", ROOT_DIR)
+# /home/gillesa/github/vcub_keeper/src
+# /home/gillesa/github/vcub_keeper
+
+
+# print("ROOT_DATA_RAW", ROOT_DATA_RAW)
+# print("ROOT_DATA_CLEAN", ROOT_DATA_CLEAN)
+# print("ROOT_DATA_REF", ROOT_DATA_REF)
+# print("ROOT_MODEL", ROOT_MODEL)
+# print("ROOT_TESTS_DATA", ROOT_TESTS_DATA)  # This will raise an error if ROOT_TESTS_DATA is not
+
+
+# /home/gillesa/github/vcub_keeper/src/vcub_keeper/data/raw/
+# /home/gillesa/github/vcub_keeper/data/raw
+
+# /home/gillesa/github/vcub_keeper/vcub_keeper/tests/data_for_tests/
+# /home/gillesa/github/vcub_keeper/tests/data_for_tests
