--- vcub_keeper/__init__.py
@@ -1,1 +1,0 @@
-__version__ = '1.2.1'
