--- vcub_keeper/tests/data_for_tests/data_test_transf_json_to_df.csv
@@ -1,145 +1,0 @@
-station_id,date,available_stands,available_bikes,status,transactions_in,transactions_out,transactions_all
-106,2023-03-10 01:10:00+01:00,32,8,1,0.0,0.0,0.0
-106,2023-03-10 01:20:00+01:00,32,8,1,0.0,0.0,0.0
-106,2023-03-10 01:30:00+01:00,32,8,1,0.0,0.0,0.0
-106,2023-03-10 01:40:00+01:00,33,8,1,0.0,1.0,0.0
-106,2023-03-10 01:50:00+01:00,33,7,1,0.0,0.0,1.0
-106,2023-03-10 02:00:00+01:00,33,7,1,0.0,0.0,0.0
-106,2023-03-10 02:10:00+01:00,33,7,1,0.0,0.0,0.0
-106,2023-03-10 02:20:00+01:00,33,7,1,0.0,0.0,0.0
-106,2023-03-10 02:30:00+01:00,33,7,1,0.0,0.0,0.0
-106,2023-03-10 02:40:00+01:00,34,7,1,0.0,1.0,0.0
-106,2023-03-10 02:50:00+01:00,35,6,1,0.0,1.0,1.0
-106,2023-03-10 03:00:00+01:00,36,5,1,0.0,1.0,1.0
-106,2023-03-10 03:10:00+01:00,35,5,1,0.0,0.0,0.0
-106,2023-03-10 03:20:00+01:00,35,5,1,0.0,0.0,0.0
-106,2023-03-10 03:30:00+01:00,38,4,1,0.0,3.0,1.0
-106,2023-03-10 03:40:00+01:00,38,2,1,0.0,0.0,2.0
-106,2023-03-10 03:50:00+01:00,38,2,1,0.0,0.0,0.0
-106,2023-03-10 04:00:00+01:00,39,2,1,0.0,1.0,0.0
-106,2023-03-10 04:10:00+01:00,39,1,1,0.0,0.0,1.0
-106,2023-03-10 04:20:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 04:30:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 04:40:00+01:00,39,2,1,1.0,0.0,1.0
-106,2023-03-10 04:50:00+01:00,39,1,1,0.0,0.0,1.0
-106,2023-03-10 05:00:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 05:10:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 05:20:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 05:30:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 05:40:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 05:50:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 06:00:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 06:10:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 06:20:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 06:30:00+01:00,40,0,1,0.0,1.0,1.0
-106,2023-03-10 06:40:00+01:00,40,0,1,0.0,0.0,0.0
-106,2023-03-10 06:50:00+01:00,40,0,1,0.0,0.0,0.0
-106,2023-03-10 07:00:00+01:00,40,0,1,0.0,0.0,0.0
-106,2023-03-10 07:10:00+01:00,40,0,1,0.0,0.0,0.0
-106,2023-03-10 07:20:00+01:00,40,0,1,0.0,0.0,0.0
-106,2023-03-10 07:30:00+01:00,40,0,1,0.0,0.0,0.0
-106,2023-03-10 07:40:00+01:00,40,1,1,1.0,0.0,1.0
-106,2023-03-10 07:50:00+01:00,39,1,1,0.0,0.0,0.0
-106,2023-03-10 08:00:00+01:00,38,4,1,3.0,0.0,3.0
-106,2023-03-10 08:10:00+01:00,37,3,1,0.0,0.0,1.0
-106,2023-03-10 08:20:00+01:00,38,3,1,0.0,1.0,0.0
-106,2023-03-10 08:30:00+01:00,37,3,1,0.0,0.0,0.0
-106,2023-03-10 08:40:00+01:00,37,5,1,2.0,0.0,2.0
-106,2023-03-10 08:50:00+01:00,35,5,1,0.0,0.0,0.0
-106,2023-03-10 09:00:00+01:00,26,15,1,10.0,1.0,10.0
-106,2023-03-10 09:10:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 09:20:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 09:30:00+01:00,26,14,1,0.0,1.0,1.0
-106,2023-03-10 09:40:00+01:00,26,15,1,1.0,0.0,1.0
-106,2023-03-10 09:50:00+01:00,25,16,1,1.0,0.0,1.0
-106,2023-03-10 10:00:00+01:00,25,15,1,0.0,0.0,1.0
-106,2023-03-10 10:10:00+01:00,24,16,1,1.0,0.0,1.0
-106,2023-03-10 10:20:00+01:00,23,17,1,1.0,0.0,1.0
-106,2023-03-10 10:30:00+01:00,25,15,1,0.0,2.0,2.0
-106,2023-03-10 10:40:00+01:00,24,16,1,1.0,0.0,1.0
-106,2023-03-10 10:50:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 11:00:00+01:00,22,18,1,2.0,0.0,2.0
-106,2023-03-10 11:10:00+01:00,22,18,1,0.0,0.0,0.0
-106,2023-03-10 11:20:00+01:00,22,18,1,0.0,0.0,0.0
-106,2023-03-10 11:30:00+01:00,22,18,1,0.0,0.0,0.0
-106,2023-03-10 11:40:00+01:00,23,18,1,0.0,1.0,0.0
-106,2023-03-10 11:50:00+01:00,24,17,1,0.0,1.0,1.0
-106,2023-03-10 12:00:00+01:00,24,16,1,0.0,0.0,1.0
-106,2023-03-10 12:10:00+01:00,24,17,1,1.0,0.0,1.0
-106,2023-03-10 12:20:00+01:00,25,17,1,1.0,1.0,2.0
-106,2023-03-10 12:30:00+01:00,23,17,1,0.0,0.0,0.0
-106,2023-03-10 12:40:00+01:00,24,16,1,0.0,1.0,1.0
-106,2023-03-10 12:50:00+01:00,23,17,1,1.0,0.0,1.0
-106,2023-03-10 13:00:00+01:00,23,17,1,0.0,0.0,0.0
-106,2023-03-10 13:10:00+01:00,22,19,1,2.0,0.0,2.0
-106,2023-03-10 13:20:00+01:00,21,19,1,0.0,0.0,0.0
-106,2023-03-10 13:30:00+01:00,21,19,1,0.0,0.0,0.0
-106,2023-03-10 13:40:00+01:00,21,20,1,1.0,0.0,1.0
-106,2023-03-10 13:50:00+01:00,21,20,1,0.0,0.0,0.0
-106,2023-03-10 14:00:00+01:00,20,20,1,0.0,0.0,0.0
-106,2023-03-10 14:10:00+01:00,20,20,1,0.0,0.0,0.0
-106,2023-03-10 14:20:00+01:00,20,21,1,1.0,0.0,1.0
-106,2023-03-10 14:30:00+01:00,18,22,1,1.0,0.0,1.0
-106,2023-03-10 14:40:00+01:00,18,22,1,0.0,0.0,0.0
-106,2023-03-10 14:50:00+01:00,18,22,1,0.0,0.0,0.0
-106,2023-03-10 15:00:00+01:00,18,22,1,0.0,0.0,0.0
-106,2023-03-10 15:10:00+01:00,19,22,1,0.0,1.0,0.0
-106,2023-03-10 15:20:00+01:00,18,23,1,1.0,0.0,1.0
-106,2023-03-10 15:30:00+01:00,17,24,1,1.0,0.0,1.0
-106,2023-03-10 15:40:00+01:00,16,24,1,0.0,0.0,0.0
-106,2023-03-10 15:50:00+01:00,16,24,1,0.0,0.0,0.0
-106,2023-03-10 16:00:00+01:00,16,24,1,0.0,0.0,0.0
-106,2023-03-10 16:10:00+01:00,16,25,1,1.0,0.0,1.0
-106,2023-03-10 16:20:00+01:00,15,25,1,0.0,0.0,0.0
-106,2023-03-10 16:30:00+01:00,14,26,1,1.0,0.0,1.0
-106,2023-03-10 16:40:00+01:00,14,27,1,1.0,0.0,1.0
-106,2023-03-10 16:50:00+01:00,15,25,1,0.0,1.0,2.0
-106,2023-03-10 17:00:00+01:00,16,24,1,0.0,1.0,1.0
-106,2023-03-10 17:10:00+01:00,18,24,1,0.0,2.0,0.0
-106,2023-03-10 17:20:00+01:00,18,22,1,0.0,0.0,2.0
-106,2023-03-10 17:30:00+01:00,17,23,1,1.0,0.0,1.0
-106,2023-03-10 17:40:00+01:00,19,23,1,0.0,2.0,0.0
-106,2023-03-10 17:50:00+01:00,19,21,1,0.0,0.0,2.0
-106,2023-03-10 18:00:00+01:00,19,21,1,0.0,0.0,0.0
-106,2023-03-10 18:10:00+01:00,20,21,1,0.0,1.0,0.0
-106,2023-03-10 18:20:00+01:00,21,20,1,0.0,1.0,1.0
-106,2023-03-10 18:30:00+01:00,21,19,1,0.0,0.0,1.0
-106,2023-03-10 18:40:00+01:00,23,18,1,0.0,2.0,1.0
-106,2023-03-10 18:50:00+01:00,23,17,1,0.0,0.0,1.0
-106,2023-03-10 19:00:00+01:00,23,17,1,0.0,0.0,0.0
-106,2023-03-10 19:10:00+01:00,23,17,1,0.0,0.0,0.0
-106,2023-03-10 19:20:00+01:00,23,17,1,0.0,0.0,0.0
-106,2023-03-10 19:30:00+01:00,23,17,1,0.0,0.0,0.0
-106,2023-03-10 19:40:00+01:00,24,16,1,0.0,1.0,1.0
-106,2023-03-10 19:50:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 20:00:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 20:10:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 20:20:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 20:30:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 20:40:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 20:50:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 21:00:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 21:10:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 21:20:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 21:30:00+01:00,24,16,1,0.0,0.0,0.0
-106,2023-03-10 21:40:00+01:00,25,15,1,0.0,1.0,1.0
-106,2023-03-10 21:50:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 22:00:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 22:10:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 22:20:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 22:30:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 22:40:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 22:50:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 23:00:00+01:00,26,15,1,0.0,1.0,0.0
-106,2023-03-10 23:10:00+01:00,26,14,1,0.0,0.0,1.0
-106,2023-03-10 23:20:00+01:00,26,14,1,0.0,0.0,0.0
-106,2023-03-10 23:30:00+01:00,26,15,1,1.0,0.0,1.0
-106,2023-03-10 23:40:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-10 23:50:00+01:00,25,15,1,0.0,0.0,0.0
-106,2023-03-11 00:00:00+01:00,26,14,1,0.0,1.0,1.0
-106,2023-03-11 00:10:00+01:00,26,14,1,0.0,0.0,0.0
-106,2023-03-11 00:20:00+01:00,26,14,1,0.0,0.0,0.0
-106,2023-03-11 00:30:00+01:00,26,14,1,0.0,0.0,0.0
-106,2023-03-11 00:40:00+01:00,26,14,1,0.0,0.0,0.0
-106,2023-03-11 00:50:00+01:00,26,14,1,0.0,0.0,0.0
-106,2023-03-11 01:00:00+01:00,26,14,1,0.0,0.0,0.0
